change NewDeck to NewElement for both Deck and Card
