+++ frontend/src/App.tsx
@@ -14,7 +14,6 @@ function App() {
           <Route index element={<ReviewView />} />
           <Route path="/:id" element={<ReviewView />} />
           <Route path="/edit/:id" element={<EditView />}>
-            <Route path="/edit/:id/create" element={<EditView />} />
             <Route path="/edit/:id/:card_id" element={<EditView />} />
           </Route>
           <Route path="/create" element={<CreateDeckView />} />
