--- frontend/src/components/Bars/NewDeck/NewDeck.styles.tsx
@@ -1,7 +1,0 @@
-const useStyles = () => {
-    return {
-        newDeck: "bg-neutral-200 grow hover:shadow-md active:bg-amber-300 transition-shadow duration-300 hover:border-2 hover:border-amber-300 p-4 flex flex-col items-center justify-center w-full h-38 cursor-pointer rounded-lg",
-    };
-};
-
-export default useStyles;
