+++ frontend/src/components/Bars/NewElement/NewElementProps.types.ts
@@ -1,0 +1,3 @@
+export interface NewDeckProps {
+    navigate: () => void,
+}
