+++ frontend/src/components/EditView/CardsList/CardListProps.types.ts
@@ -3,5 +3,6 @@ import type { CardProps } from "../Card/CardProps.types";
 export interface CardListProps {
     cards: CardProps[],
     chooseCardToEdit: (cardId: number) => void,
-    deleteCard: (cardId: number) => void
+    deleteCard: (cardId: number) => void,
+    addCard: () => void,
 }
